requirements.txt
Flask==2.3.3
gunicorn==21.2.0
pandas
numpy
scikit-learn
google-generativeai
python-dotenv
